fixed issues with the file chooser. fixed an issue where the image render component was not scaling the image for the case where it isn't a sprite sheet.

+++ Source/Kinectitude/Editor/Models/Properties/Property.cs
@@ -111,10 +111,14 @@ namespace Kinectitude.Editor.Models.Properties
                     if (result == true)
                     {
                         string fileName = System.IO.Path.GetFileName(pathName);
-                        string targetPath = @Workspace.Instance.Project.Location;
+                        string localProject = System.IO.Path.Combine(Workspace.Instance.Project.Location, Workspace.Instance.Project.GameRoot);
+                        string targetPath = @localProject;
                         string destFile = System.IO.Path.Combine(targetPath, fileName);
 
-                        System.IO.File.Copy(pathName, destFile, true);
+                        if (!System.IO.File.Exists(destFile))
+                        {
+                            System.IO.File.Copy(pathName, destFile, true);
+                        }
 
                         Value = new Value("\"" + fileName + "\"");
                     }
